Some code refactoring to eliminate superfluous code
