+++ src/main/resources/com/kiuwan/plugins/kiuwanJenkinsPlugin/KiuwanGlobalConfigDescriptor/config.jelly
@@ -56,7 +56,7 @@
 			<j:scope>
 				<f:repeatable field="connectionProfiles" name="connectionProfiles" var="connectionProfile" minimum="0" header="Connection profile">
 					<table width="100%">
-						<st:include page="/com/kiuwan/plugins/kiuwanJenkinsPlugin/KiuwanGlobalConfigDescriptor/profile.jelly" />
+						<st:include page="/com/kiuwan/plugins/kiuwanJenkinsPlugin/KiuwanGlobalConfigDescriptor/config_profile.jelly" />
 						<f:entry>
 							<div align="right">
 								<f:repeatableDeleteButton />
